+++ CHANGELOG.md
@@ -4,7 +4,7 @@ nylas-python Changelog
 Unreleased
 ----------------
 * Add Outbox support
-* Add UAS Support (Integrations API, Grants API, UAS Hosted Authentication)
+* Add support for new (beta) Integrations authentication (Integrations API, Grants API, Hosted Authentication for Integrations)
 * Add support for `limit` and `offset` for message/thread search
 * Enable Nylas API v2.5 support
 * Fix `Draft` not sending metadata